refactor: move quote_value to progress.py (Issue #296)
quote_value is a neutral log-field quoting helper, unrelated to Pi
session JSONL parsing. Move it to progress.py and update every importer
(cli_source, pilot_setup, systemd_deploy, runner, pi_process) plus the
Pi activity module's own format_run_scene so semantically unrelated
modules no longer import pi_activity for it.
